little variant of FB's op command.
Here the list of associations extension <-> application is given in a file as a list of CSVs. This file ils in the user's $HOME directory. If the file does not exist, it is created with a list of pairs.

The system can therefore be configurated by editing this $HOME/.oprc file and adding/changing some pairs.